# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Einstein API",
        "description": "API para gerenciamento de clientes e endereços",
        "contact": {
            "email": "contato@einstein.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desenvolvimento"
        }
    ],
    "paths": {
        "/customers/{customerId}/addresses": {
            "get": {
                "tags": [
                    "Customer Addresses"
                ],
                "summary": "Listar todos os endereços de um cliente",
                "operationId": "9def6e6177e81b56e16a15c189c76d78",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de endereços do cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Address"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customer Addresses"
                ],
                "summary": "Criar um novo endereço para um cliente",
                "operationId": "ad1999d406d892a7f98b5870c18255e2",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Address"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Endereço criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Endereço criado com sucesso"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Address"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/customers/{customerId}/addresses/{id}": {
            "get": {
                "tags": [
                    "Customer Addresses"
                ],
                "summary": "Exibir um endereço específico de um cliente",
                "operationId": "0b3bb6a72722a58455c7e7acf0bb9f6d",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Address"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customer Addresses"
                ],
                "summary": "Deletar um endereço específico de um cliente",
                "operationId": "49bdbc3fc20736cc00e0fa8f91f3ee09",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Endereço deletado com sucesso"
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Customer Addresses"
                ],
                "summary": "Atualizar um endereço específico de um cliente",
                "operationId": "1ff6aae1e5785012d50176b3c9a17fda",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Address"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Endereço atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Endereço atualizado com sucesso"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Address"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Address": {
                "title": "Address",
                "required": [
                    "customer_id",
                    "address",
                    "number",
                    "zip_code"
                ],
                "properties": {
                    "customer_id": {
                        "description": "ID do cliente",
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "description": "Endereço",
                        "type": "string",
                        "example": "Rua das Flores"
                    },
                    "number": {
                        "description": "Número",
                        "type": "string",
                        "example": "123"
                    },
                    "complement": {
                        "description": "Complemento",
                        "type": "string",
                        "example": "Apto 456"
                    },
                    "zip_code": {
                        "description": "CEP",
                        "type": "string",
                        "example": "12345-678"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}